requirements.txt
Flask
google-generativeai
openai
